+++ android/src/main/java/com/artinoise/recorder/FluidSynthDriver.java
@@ -1,4 +1,4 @@
-package org.billthefarmer.mididriver;
+package com.artinoise.recorder;
 import android.util.Log;
