+++ android/src/main/java/com/artinoise/recorder/GeneralMidiConstants.java
@@ -20,7 +20,7 @@
 //
 ///////////////////////////////////////////////////////////////////////////////
 
-package org.billthefarmer.mididriver;
+package com.artinoise.recorder;
 
 /**
  * General midi instrument constants
